refactor(cli): consolidate manual helpers
