+++ src/cli/commands/generate-gml-identifiers.js
@@ -10,7 +10,7 @@ import {
     toNormalizedLowerCaseSet,
     toPosixPath
 } from "../lib/shared-deps.js";
-import { writeManualJsonArtifact } from "../lib/manual-file-helpers.js";
+import { writeManualJsonArtifact } from "../lib/manual/file-helpers.js";
 import {
     DEFAULT_MANUAL_REPO,
     buildManualRepositoryEndpoints,
@@ -29,18 +29,18 @@ import {
 import {
     applyManualEnvOptionOverrides,
     IDENTIFIER_VM_TIMEOUT_ENV_VAR
-} from "../lib/manual-env.js";
+} from "../lib/manual/environment.js";
 import { applyStandardCommandOptions } from "../lib/command-standard-options.js";
 import {
     applySharedManualCommandOptions,
     resolveManualCommandOptions
 } from "../lib/manual/command-options.js";
 import { wrapInvalidArgumentResolver } from "../lib/command-parsing.js";
-import { createManualAccessContext } from "../lib/manual-command-context.js";
+import { createManualAccessContext } from "../lib/manual/context.js";
 import {
     decodeManualKeywordsPayload,
     decodeManualTagsPayload
-} from "../lib/manual-payload-validation.js";
+} from "../lib/manual/payload-validation.js";
 
 const {
     environment: {
